renamed RequestHash to RequestRoot
