+++ lib/request_root.rb
@@ -1,5 +1,5 @@
 require 'pry'
-class RequestHash
+class RequestRoot
   attr_reader :request_root, :request_lines
   def initialize(request_lines)
     @request_lines = request_lines
